Move commands to their own files 🚚
