+++ src/commands/appSettings/addAppSetting.ts
@@ -1,0 +1,16 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { AppSettingsTreeItem } from "vscode-azureappservice";
+import { IActionContext } from "vscode-azureextensionui";
+import { ext } from "../../extensionVariables";
+
+export async function addAppSetting(context: IActionContext, node?: AppSettingsTreeItem): Promise<void> {
+    if (!node) {
+        node = await ext.tree.showTreeItemPicker<AppSettingsTreeItem>(AppSettingsTreeItem.contextValue, context);
+    }
+
+    await node.createChild(context);
+}
