+++ src/commands/startWebApp.ts
@@ -1,0 +1,26 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { SiteClient } from "vscode-azureappservice";
+import { IActionContext } from "vscode-azureextensionui";
+import { SiteTreeItem } from "../explorer/SiteTreeItem";
+import { WebAppTreeItem } from "../explorer/WebAppTreeItem";
+import { ext } from "../extensionVariables";
+
+export async function startWebApp(context: IActionContext, node?: SiteTreeItem): Promise<void> {
+    if (!node) {
+        node = await ext.tree.showTreeItemPicker<WebAppTreeItem>(WebAppTreeItem.contextValue, context);
+    }
+
+    const client: SiteClient = node.root.client;
+    const startingApp: string = `Starting "${client.fullName}"...`;
+    const startedApp: string = `"${client.fullName}" has been started.`;
+
+    await node.runWithTemporaryDescription("Starting...", async () => {
+        ext.outputChannel.appendLog(startingApp);
+        await client.start();
+        ext.outputChannel.appendLog(startedApp);
+    });
+}
