+++ src/explorer/WebAppTreeItem.ts
@@ -3,16 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ResourceManagementClient } from 'azure-arm-resource';
 import { AppServicePlan } from 'azure-arm-website/lib/models';
-import * as fs from 'fs-extra';
-import * as path from 'path';
-import * as vscode from 'vscode';
 import { ISiteTreeRoot } from 'vscode-azureappservice';
-import { AzExtTreeItem, AzureTreeItem, createAzureClient, TreeItemIconPath } from 'vscode-azureextensionui';
+import { AzExtTreeItem, AzureTreeItem, TreeItemIconPath } from 'vscode-azureextensionui';
 import { ext } from '../extensionVariables';
 import { nonNullProp, nonNullValue } from '../utils/nonNull';
-import { getIconPath, getResourcesPath } from '../utils/pathUtils';
+import { getIconPath } from '../utils/pathUtils';
 import { DeploymentSlotsNATreeItem, DeploymentSlotsTreeItem } from './DeploymentSlotsTreeItem';
 import { DeploymentSlotTreeItem } from './DeploymentSlotTreeItem';
 import { SiteTreeItem } from './SiteTreeItem';
@@ -56,70 +52,5 @@ export class WebAppTreeItem extends SiteTreeItem {
         }
 
         return super.pickTreeItemImpl(expectedContextValues);
-    }
-
-    public async generateDeploymentScript(): Promise<void> {
-        const resourceClient: ResourceManagementClient = createAzureClient(this.root, ResourceManagementClient);
-        const tasks = Promise.all([
-            resourceClient.resourceGroups.get(this.root.client.resourceGroup),
-            this.root.client.getAppServicePlan(),
-            this.root.client.getSiteConfig(),
-            this.root.client.listApplicationSettings()
-        ]);
-
-        const taskResults = await tasks;
-        const rg = nonNullValue(taskResults[0]);
-        const plan = taskResults[1];
-        const siteConfig = nonNullValue(taskResults[2]);
-        const appSettings = nonNullValue(taskResults[3]);
-
-        let script: string;
-
-        if (!siteConfig.linuxFxVersion) {
-            const scriptTemplate = await this.loadScriptTemplate('windows-default.sh');
-            script = scriptTemplate;
-        } else if (siteConfig.linuxFxVersion.toLowerCase().startsWith('docker')) {
-            const scriptTemplate = await this.loadScriptTemplate('docker-image.sh');
-            let serverUrl: string | undefined;
-            let serverUser: string | undefined;
-            let serverPwd: string | undefined;
-            if (appSettings.properties) {
-                serverUrl = appSettings.properties.DOCKER_REGISTRY_SERVER_URL;
-                serverUser = appSettings.properties.DOCKER_REGISTRY_SERVER_USERNAME;
-                serverPwd = appSettings.properties.DOCKER_REGISTRY_SERVER_PASSWORD;
-            }
-            const containerParameters =
-                (serverUrl ? `SERVERURL="${serverUrl}"\n` : '') +
-                (serverUser ? `SERVERUSER="${serverUser}"\n` : '') +
-                (serverPwd ? `SERVERPASSWORD="*****"\n` : '');
-            const containerCmdParameters =
-                (serverUrl ? '--docker-registry-server-url $SERVERURL ' : '') +
-                (serverUser ? '--docker-registry-server-user $SERVERUSER ' : '') +
-                (serverPwd ? '--docker-registry-server-password $SERVERPASSWORD ' : '');
-            script = scriptTemplate.replace('%RUNTIME%', siteConfig.linuxFxVersion)
-                .replace('%IMAGENAME%', siteConfig.linuxFxVersion.substring(siteConfig.linuxFxVersion.indexOf('|') + 1))
-                .replace('%DOCKER_PARA%', containerParameters)
-                .replace('%CTN_CMD_PARA%', containerCmdParameters);
-        } else {    // Stock linux image
-            const scriptTemplate = await this.loadScriptTemplate('linux-default.sh');
-            script = scriptTemplate.replace('%RUNTIME%', siteConfig.linuxFxVersion);
-        }
-
-        // tslint:disable:no-non-null-assertion
-        script = script.replace('%SUBSCRIPTION_NAME%', this.root.subscriptionDisplayName)
-            .replace('%RG_NAME%', rg.name!)
-            .replace('%LOCATION%', rg.location)
-            .replace('%PLAN_NAME%', plan!.name!)
-            .replace('%PLAN_SKU%', plan!.sku!.name!)
-            .replace('%SITE_NAME%', this.root.client.siteName);
-        // tslint:enable:no-non-null-assertion
-
-        const doc = await vscode.workspace.openTextDocument({ language: 'shellscript', content: script });
-        await vscode.window.showTextDocument(doc);
-    }
-
-    private async loadScriptTemplate(scriptName: string): Promise<string> {
-        const templatePath = path.join(getResourcesPath(), 'deploymentScripts', scriptName);
-        return await fs.readFile(templatePath, 'utf8');
     }
 }
